Move az selection component into single place
